refactor: use separate files for package_*
